Pass the indexer logger into the block processor.
