+++ util/util.go
@@ -7,9 +7,6 @@ import (
 	"unicode"
 	"unicode/utf8"
 
-	"github.com/algorand/go-algorand/data/basics"
-	"github.com/algorand/go-algorand/data/bookkeeping"
-	"github.com/algorand/go-algorand/ledger/ledgercore"
 	"github.com/algorand/go-codec/codec"
 )
 
@@ -54,33 +51,6 @@ func JSONOneLine(obj interface{}) string {
 	enc := codec.NewEncoderBytes(&b, oneLineJSONCodecHandle)
 	enc.MustEncode(obj)
 	return string(b)
-}
-
-// CreateInitState makes an initState
-func CreateInitState(genesis *bookkeeping.Genesis) (ledgercore.InitState, error) {
-	balances, err := genesis.Balances()
-	if err != nil {
-		return ledgercore.InitState{}, fmt.Errorf("MakeProcessor() err: %w", err)
-	}
-	genesisBlock, err := bookkeeping.MakeGenesisBlock(genesis.Proto, balances, genesis.ID(), genesis.Hash())
-	if err != nil {
-		return ledgercore.InitState{}, fmt.Errorf("MakeProcessor() err: %w", err)
-	}
-
-	accounts := make(map[basics.Address]basics.AccountData)
-	for _, alloc := range genesis.Allocation {
-		address, err := basics.UnmarshalChecksumAddress(alloc.Address)
-		if err != nil {
-			return ledgercore.InitState{}, fmt.Errorf("openLedger() decode address err: %w", err)
-		}
-		accounts[address] = alloc.State
-	}
-	initState := ledgercore.InitState{
-		Block:       genesisBlock,
-		Accounts:    accounts,
-		GenesisHash: genesisBlock.GenesisHash(),
-	}
-	return initState, nil
 }
 
 func init() {
